requirements.txt
docling==2.61.2
pandas==2.3.3
marker-pdf[full]==1.10.1
camelot-py[base]==1.0.9
pytesseract==0.3.13
# pillow==12.0.0py